Move the CAI temporary table to a local sqlite database.
* This reduces network load by 50% by reducing round trips to the
cloudsql server, removing 700MB of network traffic when 350MB of test
assets is imported.
* Total time to import data into the sqlite database is further reduced
from 150 seconds to 30 seconds for 350MB of test assets.
* Temporary file is cleaned up at the end of inventory, reducing storage
requirements on server.
